refactor: standardize naming for command and event handler files in PlaceType module
