+++ apps/api/src/modules/parking/application/command-handlers/delete-place-type-command-handler.ts
