+++ apps/api/src/modules/parking/application/command-handlers/update-place-type-command-handler.ts
