+++ apps/api/src/modules/parking/application/event-handlers/index.ts
@@ -1,4 +1,4 @@
-import { PlaceTypeDeletedEventHandler } from './place-type-deleted-event.handler';
+import { PlaceTypeDeletedEventHandler } from './place-type-deleted.event-handler';
 import { ParkingAddonDeletedEventHandler } from './parking-addon-deleted.event-handler';
 
 export const eventHandlers = [
