+++ apps/api/src/modules/parking/application/event-handlers/place-type-deleted.event-handler.ts
